Refactoring codegen
