+++ wasm/src/lib.rs
@@ -83,7 +83,7 @@ pub fn get_qir(
     let profile =
         Profile::from_str(profile).map_err(|()| format!("Invalid target profile {profile}"))?;
     if language_features.contains(LanguageFeatures::PreviewQirGen) {
-        _get_qir_preview(sources, language_features, profile)
+        qsc::codegen::get_qir(sources, language_features, profile.into())
     } else {
         _get_qir(sources, language_features)
     }
@@ -115,53 +115,6 @@ fn _get_qir(sources: SourceMap, language_features: LanguageFeatures) -> Result<S
     let package = store.insert(unit);
 
     generate_qir(&store, package).map_err(|e| e.0.to_string())
-}
-
-fn _get_qir_preview(
-    sources: SourceMap,
-    language_features: LanguageFeatures,
-    profile: Profile,
-) -> Result<String, String> {
-    let core = compile::core();
-    let mut package_store = PackageStore::new(core);
-    let std = compile::std(&package_store, profile.into());
-    let std = package_store.insert(std);
-
-    let (unit, errors) = qsc::compile::compile(
-        &package_store,
-        &[std],
-        sources,
-        PackageType::Exe,
-        profile.into(),
-        language_features,
-    );
-
-    // Ensure it compiles before trying to add it to the store.
-    if !errors.is_empty() {
-        // This should never happen, as the program should be checked for errors before trying to
-        // generate code for it. But just in case, simply report the failure.
-        return Err("Failed to generate QIR".to_string());
-    }
-
-    let package_id = package_store.insert(unit);
-
-    let caps_results =
-        PassContext::run_fir_passes_on_hir(&package_store, package_id, profile.into());
-    // Ensure it compiles before trying to add it to the store.
-    match caps_results {
-        Ok(compute_properties) => hir_to_qir(
-            &package_store,
-            package_id,
-            profile.into(),
-            Some(compute_properties),
-        )
-        .map_err(|e| e.to_string()),
-        Err(_) => {
-            // This should never happen, as the program should be checked for errors before trying to
-            // generate code for it. But just in case, simply report the failure.
-            Err("Failed to generate QIR".to_string())
-        }
-    }
 }
 
 #[wasm_bindgen]
